fixed bug causing crashes during the deployment on 20080123
Reason: manual transaction databases require to lock tables before writing. The function
generating the sql 'LOCK TABLES' command was buggy, and thus only generated the command
'LOCK TABLES', without any tables listed, which is a faulty mysql command.

There was a gcc warning in that piece of code earlier, that's why I've changed it in
the first place. It was even buggy: g_slist_append *can* change the beginning of the list,
that's why it returns a pointer, and gives a warning if that pointer is left unused.

This fix now probably fixed a lot other mal-behaviours as well (memleak and operation on
non-locked tables).