- added dummification support to text_export - fixed bug causing nrtm problems
NRTM updates database in a different way: when recreating a previously
deleted pkey, it simply keeps the object_id, and updates (deletes) the
sequence_id=0 record in the last table to the following sequence_id, puts
the object in there, etc..., so it basically overwrites the deletion
placeholder record in the last table, keeping the old object_id.

It's funny as this proves that NRTM is NOT a proper replication - luckily
this bug keeps the serials the same, but the sql DB itself is vastly
different.

As a workaround, I've changed the inner joins to left join. Of course this
will result in timestamp being NULL in NRTM delete+re-create cases, but that
data is lost anyway (as the placeholder last record is overwritten).