text_export also creates a CURRENTSERIAL file.
The old way of creating the CURRENTSERIAL file was to execute '-q sources'
before the sql dump (followed by the split file dump). Since updates were
enabled, this resulted in incorrect CURRENTSERIAL, which is otherwise
essential to correctly start the NRTM stream.

Since we don't provide sql dumps on the ftp site anymore, I've changed
text_export to output the highest serial_id (which is the last thanks to the
auto_increment column in serials table). This is then published on the
ftp site instead, so it is always in sync with the published split files.