Fixed long-standing serious bug. For some reason it didn't result in a crash on 32-bit systems.
g_hash_table_insert doesn't make a copy of thethe key and value, but relies on their content not changing.
Still, it it UT_free()d, causing the g_str_equal in the following check to fail.

Another great example of crappy c code only working on a very-very-very specific environment, because of
ignoring standards.