fixed bug of die()ing inside the CA module.
The do_nice_die() called a function inside the CA module, which had a lock
inside. If CA module died while that lock was still held, do_nice_die (which
is also the segfault handler) got in a deadlock, and the whole process
froze. Solution was to read the config value at startup, store it in a
variable, and reference that variable instead of calling CA module for it.

This is a bit hackish, but then again, this is a variable that won't change
during the lifetime of whois-server.